packages/naaviq-server/pyproject.toml
[project]
name = "naaviq-server"
version = "0.1.0"
description = "FastAPI server for Naaviq — ties all packages together"
requires-python = ">=3.12"
dependencies = [
    "naaviq-core>=0.1.0",
    "naaviq-graph>=0.1.0",
    "naaviq-voice>=0.1.0",
    "naaviq-rag>=0.1.0",
    "naaviq-tools[calendar]>=0.1.0",
    "naaviq-channels>=0.1.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    "python-jose[cryptography]>=3.3.0",
    "pwdlib[argon2]>=0.2.0",
    "cryptography>=42.0.0",
    "pydantic[email]>=2.7.0",
    "pydantic-settings>=2.3.0",
    "slowapi>=0.1.9",
    "sentry-sdk[fastapi]>=2.0.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "structlog>=24.0.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "langgraph-checkpoint-postgres>=2.0",
    "psycopg[binary]>=3.1.0",
    "psycopg-pool>=3.2.0",
    # LiveKit server SDK — used to create rooms + generate access tokens
    "livekit-api>=1.0",
    # Twilio — webhook signature verification
    "twilio>=9.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "httpx>=0.27", "ruff>=0.4", "mypy>=1.10"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["naaviq"]
